moved createplant logic into provider
