+++ plant-app/src/context/Plants.tsx
@@ -1,15 +1,20 @@
 import React, { createContext, useState, useContext, useEffect } from "react";
 import { BASE_API_URL } from "../constants";
-import { Plant, PlantImage } from "../types/interfaces";
+import { ApiResponse, NewPlant, Plant, PlantImage } from "../types/interfaces";
 import { useAlert } from "./Alerts";
 import { useApi } from "../utils/api";
 import noimagePlaceholder from "../assets/200x200_image_placeholder.png";
 import { useAuth } from "./Auth";
 
+interface CreatePlantProps {
+  plant: NewPlant;
+}
+
 interface PlantContextType {
   plants: Plant[];
   isLoading: boolean;
   fetchPlants: (queryID: string) => Promise<void>;
+  createPlant: (props: CreatePlantProps) => Promise<ApiResponse<Plant>>;
   nextPlantId: number;
   forceReloadPlants: () => void;
   plantImages: Record<string, string>;
@@ -31,7 +36,12 @@ function incrementLargestId(plants: Plant[]): number {
     return plant.human_id > maxId ? plant.human_id : maxId;
   }, plants[0].human_id);
 
-  return maxHumanId + 1; // Increment after finding the max
+  return maxHumanId + 1;
+}
+
+interface CreatePlantProps {
+  callApi: (url: string, options?: RequestInit) => Promise<Response>;
+  plant: NewPlant;
 }
 
 interface PlantProviderProps {
@@ -40,7 +50,7 @@ interface PlantProviderProps {
 export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
   const [plants, setPlants] = useState<Plant[]>([]);
   const [isLoading, setIsLoading] = useState(true);
-  const [nextPlantId, setNextPlantId] = useState(0);
+  const [nextPlantId, setNextPlantId] = useState(1);
   const [lastQueryID, setLastQueryID] = useState<string | null>(null);
 
   const [plantGridIsLoading, setPlantGridIsLoading] = useState<boolean>(false);
@@ -50,7 +60,6 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
   const { callApi } = useApi();
   const { isAuthenticated } = useAuth();
 
-  // Method to fetch plants from API
   const fetchPlants = async (queryID: string, forced: boolean = false) => {
     if (!forced) {
       if (queryID === lastQueryID || !queryID) {
@@ -85,19 +94,18 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
       });
   };
 
-  // Fetch plant thumbnails for the grid
+  // Fetch plant thumbnails for the grid whenever plants change
   useEffect(() => {
     // Prevents API error on logout
     if (!isAuthenticated) {
       return;
     }
-    // Check if image data for all plants already exists
+    // No need to fetch images if there are no plants
     if (!plants || plants.length === 0) {
-      return; // No need to fetch data
+      return;
     }
     setPlantGridIsLoading(true);
     const plantIds = plants.map((plant) => plant.plant_id);
-    console.log(plantIds);
     callApi(BASE_API_URL + "/images/plants/most_recent/", {
       method: "POST",
       headers: {
@@ -135,6 +143,30 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
   };
 
   // TODO: centralize all the plant create, update, delete methods here
+  const createPlant = async ({
+    plant,
+  }: CreatePlantProps): Promise<ApiResponse<Plant>> => {
+    const response = await callApi(`${BASE_API_URL}/plants/create`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(plant),
+    });
+    if (!response.ok) {
+      const error_response = await response.json();
+      const error_message = JSON.stringify(error_response);
+      return {
+        success: false,
+        data: null,
+        error: `Error: ${error_message}`,
+      };
+    }
+    return {
+      success: true,
+      data: await response.json(),
+    };
+  };
 
   return (
     <PlantContext.Provider
@@ -142,6 +174,7 @@ export const PlantProvider: React.FC<PlantProviderProps> = ({ children }) => {
         plants,
         isLoading,
         fetchPlants,
+        createPlant,
         nextPlantId,
         forceReloadPlants,
         plantImages,
