+++ plant-app/src/types/custom.d.ts
@@ -3,5 +3,3 @@ declare module "*.jpeg";
 declare module "*.png";
 declare module "*.gif";
 declare module "*.svg";
-
-declare module "react-lazy-load-image-component";
